requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.26.0
googletrans==3.1.0a0
lyricsgenius==3.0.1
python-multipart==0.0.9
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-mock==3.12.0 